Update README.md
Effective surveillance on the long-term public health impact due to war and terrorist attacks remains limited. Such health issues are commonly under-reported, specifically for a large group of individuals. For this purpose, efficient estimation of the size or undercount of the population under the risk of physical and mental health hazards is of utmost necessity. A novel trivariate Bernoulli model is developed allowing heterogeneity among the individuals and dependence between the sources of information, and an estimation methodology using a Monte Carlo-based EM algorithm is proposed. Simulation results show the superiority of the performance of the proposed method over existing competitors and robustness under model mis-specifications. The method is applied to analyse two real case studies on monitoring Amyotrophic Lateral Sclerosis (ALS) cases for the Gulf War veterans and the 9/11 terrorist attack survivors at the World Trade Center, US. The average annual cumulative incidence rate for ALS disease increases by 33% and 16% for deployed and no-deployed military personnel, respectively, after adjusting the undercount. The number of individuals exposed to the risk of physical and mental health effects due to WTC terrorist attacks increased by 42%. These results provide interesting insights that can assist in effective decision-making and policy formulation for monitoring the health status of post-war survivors.